Added extension for UIView
