+++ ReviewMonitor/AppDelegate.swift
@@ -121,16 +121,3 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         print(Date().timeIntervalSince(startBack))
     }
 }
-
-extension UIView {
-    public func addBorder(_ width: CGFloat, color: UIColor) {
-        layer.borderColor = color.cgColor
-        layer.borderWidth = width
-    }
-
-    public func cornerRadius(_ radius: CGFloat) {
-        layoutIfNeeded()
-        layer.cornerRadius = radius
-        clipsToBounds = true
-    }
-}
